optimized algorithm

+++ src/de/skipgraph/operations/SetContactOperation.java
@@ -11,16 +11,23 @@ public class SetContactOperation extends ModifyContactsOperation {
 	private ContactType contactType;
 	private final Contact contact;
 	private int level;
-	private byte prefix;
+	private int prefix;
 
-	public SetContactOperation(int level, byte prefix, ContactType contactType, Contact contact) {
+	public SetContactOperation(int level, int prefix, ContactType contactType, Contact contact) {
 		this.level = level;
 		this.prefix = prefix;
 		this.contactType = contactType;
 		this.contact = contact;
 	}
 
-	public SetContactOperation(int level, byte prefix, ContactType contactType, Node node) {
+	/**
+	 *
+	 * @param level        number of level
+	 * @param prefix       prefix for this level (usually 0 or 1)
+	 * @param contactType  PREV to update prevContact or NEXT to update nextContact
+	 * @param node         node you want to set the contact to
+	 */
+	public SetContactOperation(int level, int prefix, ContactType contactType, Node node) {
 		this.level = level;
 		this.prefix = prefix;
 		this.contactType = contactType;
